consolidate db model files to schema/ dir
